Splits up first+middle/last names and implements filtering them.
Now whenever an individual is saved to the database the save() function will handle splitting up the name. In the future we may want to have it switch them from UPPERCASE to Title Case while we're at it.

I had assumed that providing filtering against names would slow down the browsable front end but it doesn't seem to build a select, so there isn't a performance hit.