Hard codes database configuration (except password).
For some reason the os.environ RDS_... variables weren't automatically configured when I associated the RDS database with the EB environment, and adding them through the console GUI didn't seem to be working either, so I'm trying it the old fashioned way.

The password will need to be hand edited on the server.